# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SocialMetaCollector API - Collect metadata from YouTube, Instagram, and Facebook
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /youtube/videos:
    get:
      operationId: listYoutubeVideos
      tags: [youtube]
      summary: List YouTube videos with metadata
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YoutubeVideoList"
  /youtube/videos/{videoId}:
    get:
      operationId: getYoutubeVideo
      tags: [youtube]
      summary: Get specific YouTube video metadata
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Video metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YoutubeVideo"
...
components:
  schemas:
...
    YoutubeVideo:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        viewCount:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        duration:
          type: string
        thumbnailUrl:
          type: string
        publishedAt:
          type: string
          format: date-time
        platform:
          type: string
      required:
        - id
        - title
        - viewCount
        - likeCount
        - commentCount
    YoutubeVideoList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/YoutubeVideo"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
...